♻️ refactor(config): delete dead code, move single-consumer helpers home, split files by concept
internal/config had become a dumping ground: process/env configuration,
DB-backed domain types, model-resolution logic, and version-derived
helpers all shared one package. Three changes, no behavior change:

- Delete dead code: cli_plugin.go (CLITool trio), ModelConfig/
  ModelConfigToAI, and DBPath — zero references anywhere.
- Move single-consumer helpers to their consumer: the sandbox image
  tag resolver to internal/agent/sandbox (unexported), the default
  model policy to internal/store (unexported).
- Split remaining files so each holds one concept and a future package
  split is mechanical: store.go now carries only the Store interface
  (row types moved to provider.go/agent.go/channel.go), config.go
  renamed to settings.go, the STELLA_SANDBOX_BACKEND read extracted to
  sandbox_env.go, ModelTier consts folded into snapshot.go, and doc.go
  documents the two groups. Tests split to match.

Refs #701
