+++ internal/config/channel.go
@@ -1,0 +1,11 @@
+package config
+
+// Channel represents a platform channel configuration.
+type Channel struct {
+	ID      string `json:"id"`
+	Name    string `json:"name"`
+	Type    string `json:"type"`
+	AgentID string `json:"agent_id,omitempty"`
+	Enabled bool   `json:"enabled"`
+	Config  string `json:"config"`
+}
