+++ internal/config/settings_test.go
@@ -1,0 +1,25 @@
+package config
+
+import "testing"
+
+func TestSchedulerIsEnabled(t *testing.T) {
+	tr := true
+	fa := false
+	tests := []struct {
+		name    string
+		enabled *bool
+		want    bool
+	}{
+		{"nil defaults to true", nil, true},
+		{"explicit true", &tr, true},
+		{"explicit false", &fa, false},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			c := SchedulerConfig{Enabled: tt.enabled}
+			if got := c.IsEnabled(); got != tt.want {
+				t.Errorf("IsEnabled() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
